Cleanup for MacOS target.

+++ Core/Source/NSData+DTBase64.m
@@ -101,7 +101,7 @@ void *NewDTBase64Decode(
 		size_t accumulateIndex = 0;
 		while (i < length)
 		{
-			unsigned char decode = base64DecodeLookup[inputBuffer[i++]];
+			unsigned char decode = base64DecodeLookup[(int) inputBuffer[i++]];
 			if (decode != xx)
 			{
 				accumulated[accumulateIndex] = decode;
